refactor: make routes import functions
